reddit_market_research/requirements.txt
# Core Web Framework
streamlit>=1.28.0
streamlit-pdf-viewer>=0.0.12

# AI and LLM
openai>=1.0.0
vllm>=0.2.0

# Workflow Orchestration
langgraph>=0.0.40
langchain>=0.1.0

# Vector Database
qdrant-client>=1.6.0

# Reddit API
praw>=7.7.0
prawcore>=2.3.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0
pydantic>=2.0.0

# Natural Language Processing
nltk>=3.8.0

# Visualization and Reporting
matplotlib>=3.6.0
wordcloud>=1.9.0
weasyprint>=60.0

# Web and HTTP
requests>=2.28.0
aiohttp>=3.8.0
httpx>=0.24.0

# Utility Libraries
python-dotenv>=1.0.0
typing-extensions>=4.5.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Additional Dependencies for PDF Generation
markdown>=3.4.0
cssselect2>=0.7.0
html5lib>=1.1
pyphen>=0.14.0
tinycss2>=1.2.0
cffi>=1.15.0 